% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationSpec.R
\name{completeSimulation}
\alias{completeSimulation}
\title{Complete a simulation specification from a model}
\usage{
completeSimulation(model, sim)
}
\arguments{
\item{model}{A NIMBLE model.}

\item{sim}{A \code{cppp_simulation} object.}
}
\value{
A completed \code{cppp_simulation} object.
}
\description{
Fill in defaults for a simulation specification using information from a
NIMBLE model.
}
\keyword{internal}
