% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrepancySpec.R
\name{standardizeDiscrepancies}
\alias{standardizeDiscrepancies}
\title{Standardize discrepancy input}
\usage{
standardizeDiscrepancies(x)
}
\arguments{
\item{x}{A \code{cppp_discrepancy} object or a list of such objects.}
}
\value{
A list of \code{cppp_discrepancy} objects.
}
\description{
Accept either one discrepancy specification or a list of them and always
return a list of discrepancy specifications.
}
\keyword{internal}
