% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppp-package.R
\docType{package}
\name{cppp-package}
\alias{cppp}
\alias{cppp-package}
\title{cppp: Posterior Predictive Checks and Calibrated P-values}
\description{
Tools for performing posterior predictive checks and computing calibrated posterior predictive p-values (cPPP) in Bayesian models. Includes infrastructure for calibration loops, diagnostic summaries, and NIMBLE-based examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/salleuska/cpppPrototype}
  \item Report bugs at \url{https://github.com/salleuska/cpppPrototype/issues}
}

}
\author{
\strong{Maintainer}: Sally Paganin \email{paganin.1@osu.edu}

Authors:
\itemize{
  \item Perry de Valpine
}

}
\keyword{internal}
